minor fixes for internal api
* provide all table names, also prefix table names with table
* name all operations with op
* ovnnbimp.go to smaller parts
